frontend: Refactor ApartmentSalePage
- Simplify imports
- Reorganise filenames, directory structure and code
- Split code from larger components into their own component
- There shouldn't be much if any functional changes
