+++ frontend/src/features/apartment/ApartmentNewSalePage/ApartmentSalePage.tsx
@@ -3,9 +3,9 @@ import {useGetApartmentDetailQuery} from "../../../app/services";
 import {Heading, QueryStateHandler} from "../../../common/components";
 import {IApartmentDetails} from "../../../common/schemas";
 import ApartmentHeader from "../components/ApartmentHeader";
-import LoadedApartmentSalesPage from "./LoadedApartmentSalesPage";
+import LoadedApartmentSalePage from "./LoadedApartmentSalePage";
 
-const ApartmentNewSalePage = () => {
+const ApartmentSalePage = () => {
     // Main Apartment New Sale container.
     // Loads the Apartment data before rendering anything.
 
@@ -17,19 +17,19 @@ const ApartmentNewSalePage = () => {
 
     return (
         <div className="view--apartment-sales">
+            <ApartmentHeader />
             <QueryStateHandler
                 data={data}
                 error={error}
                 isLoading={isLoading}
             >
-                <ApartmentHeader />
                 <Heading type="main">
                     {data && data.prices.first_purchase_date ? "Kauppatapahtuma" : "Uudiskohteen kauppa"}
                 </Heading>
-                <LoadedApartmentSalesPage apartment={data as IApartmentDetails} />
+                <LoadedApartmentSalePage apartment={data as IApartmentDetails} />
             </QueryStateHandler>
         </div>
     );
 };
 
-export default ApartmentNewSalePage;
+export default ApartmentSalePage;
