+++ frontend/src/features/apartment/ApartmentNewSalePage/fieldsets/MaximumPriceCalculationFieldSet/index.ts
@@ -1,0 +1,3 @@
+import MaximumPriceCalculationFieldSet from "./MaximumPriceCalculationFieldSet";
+
+export default MaximumPriceCalculationFieldSet;
