+++ frontend/src/features/apartment/ApartmentNewSalePage/fieldsets/index.ts
@@ -1,0 +1,5 @@
+import ApartmentCatalogPricesFieldSet from "./ApartmentCatalogPricesFieldSet";
+import ApartmentSaleFormFieldSet from "./ApartmentSaleFormFieldSet";
+import MaximumPriceCalculationFieldSet from "./MaximumPriceCalculationFieldSet";
+
+export {ApartmentCatalogPricesFieldSet, ApartmentSaleFormFieldSet, MaximumPriceCalculationFieldSet};
